Time: 11 ms (87.76%), Space: 21.7 MB (72.05%) - LeetHub

+++ 98-validate-binary-search-tree/98-validate-binary-search-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    bool helper(TreeNode* root, long long int min, long long int max)
+    bool helper(TreeNode* root, long min, long max)
     {
         if(!root) return true;
         if(root->val <= min || root->val >= max) return false;
@@ -23,6 +23,6 @@ public:
     
     bool isValidBST(TreeNode* root) 
     {
-        return helper(root, LLONG_MIN, LLONG_MAX);
+        return helper(root, LONG_MIN, LONG_MAX);
     }
 };
